RavenDB-20238 Sharding - Studio - alerts & performance hints aren't shown
